[GSW-105] 	Save the DLT-viewr version to config/registry and delete  specific values when a new minor version of the viewer is started.
Added separate parts for Major, Minor and Patch level versions in
version.h
It will now check if the minor or major version has changed (in this
order) and will offer to either reset settings, ignore the change or
exit the app.
If user chooses to reset the settings, all the settings will be cleared.
If the settings were in a file (as opposed to windows registry) and the
file is writable by the user, the file will also be deleted.