[GSW-142] Store DLT-Viewer configuration in an ini file instead of using the registry
Implemented a facade for QSettings with singleton interface.
The facade hides the construction of the QSettings. The QSettings object
will be initialized differently in windows than other platforms.
In windows it will create the settings file in the current working
directory. On all other platforms it will rely on the QT default
behaviour.
The new class implements delegate methods for only the currently needed
QSettings methods.