Enable support for High DPI screens
DLT Viewer sizes columns and other widgets based on an assumption
of a "Low-DPI" screen resolution.

On Windows systems with High DPI screens and enabled scaling,
many GUI elements are too small and the application is very hard
to use.

This patch fixes the problem by enabling Qt's High DPI support
introduced in Qt 5.6. To ensure the project still builds fine
with Qt 5.5, the change has been wrapped into an #if testing
the QT_VERSION.