Limit symbol visibility in plugins
Currently the file transfer plugin may not work if dlt-viewer was
compiled with a compiler not restricting symbol visibility. This is
caused by the generic "Form" class which is also used in other
dlt-viewer plugins.

As soon as more than one plugin provides certain symbols (in this case
it seems to be related to QObject internals) the call to
QMetaObject::connectSlotsByName within Ui_Form fails and the signals of
the widgets are not connected to the corresponding slot.

Avoid this alltogether by not exporting symbols (unless explicitly
requested) which is the same behaviour exposed by MSVC.

Change-Id: I2c3271a3867afc0496d11ad0908534d465a5b6c6

Signed-off-by: Alexander Wenzel <Alexander.AW.Wenzel@bmw.de>