Fix check box in file transfer plugin view
Currently the check box does not work at all on Linux as Qt itself takes
care of toggling the check box (and as such the check box already has
the desired value before the slot is entered).

Implement the desired callback behaviour as part of the itemChanged
signal which will be called by Qt after the model has actually been
updated.

Change-Id: I6b127fe9b07fab1ad46f7052323f49a5cd77022d

Signed-off-by: Alexander Wenzel <Alexander.AW.Wenzel@bmw.de>