qextserialport: Link IOKit & Foundation on macOS
qextserialenumerator_osx.cpp requires IOKit.framework and
Foundation.framework, so look for them and link them into the library.

Additionally, symbols from qextserialport_* are required. Since there is
no qextserialport_osx.cpp, use the _unix one for now, which allows the
build to finish and has a reasonable chance of working on macOS.

Signed-off-by: Clemens Lang <clemens.lang@bmw-carit.de>