[searchdialog] keep the cursor position when search-text is edited in the middle
Problem:
 searchTextbox (mainwindow) and lineEditText (search dialog) trigger
update on each other when one of them is changed. This in turn, triggers
an update of the original widget with setText(). Calling setText resets
the cursor position to the end of the text.

Fix:
Avoid this by blocking signals on the widget whose text is being
(programmatically) updated.

Signed-off-by: Balaji Sundararaj <balaji.sundararaj@partner.bmw.de>