Set default CMAKE_BUILD_TYPE for single configuration builds
Currently there is no build type set resulting in neither optimization
nor debug settings enabled. This applies to single configuration
generators that are the default on Linux systems.

Set to Release by default to get an optimized build. This results in
much higher performance and is therefore a good default for most
regular users.