Fix rmdir usage in Windows build script
Apparently, there are problems setting the errorlevel
variable after rmdir was called and if errorlevel is
checked afterwards it will be always reported as 0,
although there were errors when deleting the directories/files.

Executing the "rem" command on error works around this
issue. For details see https://stackoverflow.com/a/11137825

Signed-off-by: Jens Schmer <jens.schmer@garmin.com>