Set C standard to c11
This is needed, because dlt_common.c includes `<QtGlobal>` which includes
implicitly `<qtypes.h>` which contains
```
static_assert(sizeof(ptrdiff_t) == sizeof(size_t), "Weird ptrdiff_t and size_t definitions");
...
```
But `static_assert` in C is available only from c11 standard.
Though this is working with clang, but not with gcc compiler, so we set
the standard explicitly to c11.